services/notification_service/app/core/config.py
"""
Configuration for Notification Service
"""
import os
from pathlib import Path
from dotenv import load_dotenv

# Load .env from project root (4 levels up from this file)
env_path = Path(__file__).parent.parent.parent.parent.parent / ".env"
load_dotenv(dotenv_path=env_path)


class Settings:
    """Application settings"""
    APP_NAME: str = "Notification Service"
    VERSION: str = "1.0.0"
    API_V1_PREFIX: str = "/api/v1"
    
    # Telegram Bot
    TELEGRAM_BOT_TOKEN: str = os.getenv("TELEGRAM_BOT_TOKEN", "")
    
    # Celery (for async tasks)
    CELERY_BROKER_URL: str = os.getenv("CELERY_BROKER_URL", "redis://localhost:6379/0")
    CELERY_RESULT_BACKEND: str = os.getenv("CELERY_RESULT_BACKEND", "redis://localhost:6379/0")
    
    # CORS
    BACKEND_CORS_ORIGINS: list = ["*"]


settings = Settings()
