services/notification_service/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
celery==5.4.0
redis==5.2.1
python-telegram-bot==22.1
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.7.1
pydantic[email]
